Add a bypass code to allow registration outside Protospace

+++ apiserver/apiserver/api/serializers.py
@@ -405,7 +405,13 @@ class MyRegisterSerializer(RegisterSerializer):
     def custom_signup(self, request, user):
         data = request.data
 
-        if not utils.is_request_from_protospace(request):
+        if secrets.REGISTRATION_BYPASS:
+            bypass_code = data.get('bypass_code', None)
+            allow_bypass = secrets.REGISTRATION_BYPASS == bypass_code
+        else:
+            allow_bypass = False
+
+        if not allow_bypass and not utils.is_request_from_protospace(request):
             logger.info('Request not from protospace')
             user.delete()
             raise ValidationError(dict(non_field_errors='Can only register from Protospace.'))
